line mapper refactor
