+++ VSRAD.BuildToolsTests/Errors/ParserTests.cs
@@ -19,44 +19,6 @@ host.c:4:2: warning: implicitly declaring library function 'printf' with type 'i
 host.c:4:2: note: include the header<stdio.h> or explicitly provide a declaration for 'printf'
 ";
 
-        public const string ErrString = @"
-test.c:5:10: error: invalid digit 'a' in octal constant
- return 0asdfshgmgmg
-         ^
-1 error generated.
-";
-
-        public const string PpString = @"
-# 1 ""test.c""
-# 1 ""<built-in>""
-# 1 ""<command-line>""
-# 31 ""<command-line>""
-# 1 ""/usr/include/stdc-predef.h"" 1 3 4
-# 32 ""<command-line>"" 2
-# 1 ""test.c""
-
-
-int main(int argc, char** argv)
-        {
-
-            return 0asdfshgmgmg
-
-}
-
-";
-
-        public const string Preprocessed = @"
-int main() {
-    int a = 1 + 1;
-    int b = 2 + 2;
-# 16 'source.c'
-    int c = 3 + 3;
-    int d = 4 + 4;
-# 55 'source.c'
-    int f = 0xDEAD;
-}
-";
-
         [Fact]
         public void ClangErrorTest()
         {
@@ -91,20 +53,6 @@ int main() {
             Assert.Equal(2, messages[3].Column);
             Assert.Equal("host.c", messages[3].SourceFile);
             Assert.Equal(@"include the header<stdio.h> or explicitly provide a declaration for 'printf'", messages[3].Text);
-        }
-
-        [Fact]
-        public void PreprocessMapLinesTest()
-        {
-            var lineMapping = MapLines(Preprocessed);
-            Assert.Equal(new int[] { 1, 2, 3, 4, 0, 16, 17, 0, 55, 56, 57 }, lineMapping);
-        }
-
-        [Fact]
-        public void ParseErrorsWithPreprocessedTest()
-        {
-            var messages = ExtractMessages(ErrString, PpString);
-            Assert.Equal(5, messages.First().Line);
         }
     }
 }
