+++ VSRAD.BuildToolsTests/VSRAD.BuildToolsTests.csproj
@@ -52,6 +52,7 @@
     <Reference Include="System.Configuration" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Errors\LineMapperTests.cs" />
     <Compile Include="Errors\ParserTests.cs" />
     <Compile Include="RemoteBuildTaskTests.cs" />
   </ItemGroup>
